Moved calculation routines into separate files. Created dda_util.[ch] with general purpose calculation routines. (cherry-picking from commit 446386aebd3831870452fc7337dc665e7a459968) Added new source dda_util.c to Makefile. (cherry-picking from commit 78252874d6ef1abdd363343fe4069be7170b9238)
